fix delete token

+++ app/Actions/SysAdmin/User/DeleteUserAccessToken.php
@@ -22,21 +22,24 @@ class DeleteUserAccessToken extends GrpAction
     use AsAction;
     use WithAttributes;
 
-    public function handle(PersonalAccessToken $token): bool
+    public function handle(User $user, PersonalAccessToken $token): bool
     {
-        $token->delete();
-        return true;
+        if ($token->tokenable_id == $user->id && $token->tokenable_type == class_basename($user)) {
+            $token->delete();
+            return true;
+        }
+        return false;
     }
 
     public function action(User $user, PersonalAccessToken $token, array $modelData): bool
     {
         $this->initialisation($user->group, $modelData);
-        return $this->handle($token);
+        return $this->handle($user, $token);
     }
 
     public function asController(User $user, PersonalAccessToken $token, ActionRequest $request)
     {
         $this->initialisation($user->group, $request);
-        $this->handle($token);
+        $this->handle($user, $token);
     }
 }
